+++ src/Sulu/Bundle/AdminBundle/Gruntfile.js
@@ -31,15 +31,31 @@ module.exports = function(grunt) {
             build: {
                 files: [
                     //copy fonts
-                    {expand: true, cwd: 'Resources/public/js/vendor/husky/', src: ['fonts/**'], dest: 'Resources/public/dist/'},
+                    {
+                        expand: true,
+                        cwd: 'Resources/public/js/vendor/husky/',
+                        src: ['fonts/**'],
+                        dest: 'Resources/public/dist/'
+                    },
                     //only needed due to wrong path generation
                     {src: ['Resources/views/Admin/index.html.twig'], dest: 'Resources/public/index.html.twig'},
                     {src: ['Resources/views/Security/login.html.twig'], dest: 'Resources/public/login.html.twig'},
                     {expand: true, cwd: 'Resources/public', src: ['**'], dest: 'Resources/public/bundles/suluadmin/'},
                     // copy cultures
-                    {expand: true, cwd: 'Resources/public/js/vendor/globalize', src: ['cultures/**'], dest: 'Resources/public/dist/vendor/globalize'},
+                    {
+                        expand: true,
+                        cwd: 'Resources/public/js/vendor/globalize',
+                        src: ['cultures/**'],
+                        dest: 'Resources/public/dist/vendor/globalize'
+                    },
                     // copy files
-                    {expand: true, cwd: 'Resources/public/js/vendor/husky/vendor', src: ['*'], dest: 'Resources/public/dist/vendor', filter: 'isFile'}
+                    {
+                        expand: true,
+                        cwd: 'Resources/public/js/vendor/husky/vendor',
+                        src: ['*'],
+                        dest: 'Resources/public/dist/vendor',
+                        filter: 'isFile'
+                    }
 
                 ]
             },
@@ -85,7 +101,7 @@ module.exports = function(grunt) {
             }
         },
         clean: {
-            options: { force: true },
+            options: {force: true},
             dist: {
                 files: [
                     {src: ['Resources/public/dist']}
@@ -137,10 +153,16 @@ module.exports = function(grunt) {
             },
             buildMain: {
                 options: {
-                    variables: {
-                        '/js/main.js': '/dist/main.js'
-                    },
-                    prefix: ''
+                    patterns: [
+                        {
+                            match: /v=develop/,
+                            replacement: 'v=' + time.toString()
+                        },
+                        {
+                            match: /\/js\/main\.js/,
+                            replacement: '/dist/main.js'
+                        }
+                    ]
                 },
                 files: [
                     {src: ['Resources/public/dist/app.min.js'], dest: 'Resources/public/dist/app.min.js'}
@@ -148,10 +170,12 @@ module.exports = function(grunt) {
             },
             buildLogin: {
                 options: {
-                    variables: {
-                        '/js/login.js': '/dist/login.js'
-                    },
-                    prefix: ''
+                    patterns: [
+                        {
+                            match: /v=develop/,
+                            replacement: 'v=' + time.toString()
+                        }
+                    ]
                 },
                 files: [
                     {src: ['Resources/public/dist/login.min.js'], dest: 'Resources/public/dist/login.min.js'}
@@ -209,8 +233,7 @@ module.exports = function(grunt) {
         },
         cssmin: {
             // TODO: options: { banner: '<%= meta.banner %>' },
-            compress: {
-            }
+            compress: {}
         },
         compass: {
             dev: {