+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/login.js
@@ -25,7 +25,8 @@ require.config({
     ],
     exclude: [
         'husky'
-    ]
+    ],
+    urlArgs: 'v=develop'
 });
 
 define('underscore', function() {